♻️ refactor: Temp domain 관련 코드 모두 삭제
