+++ src/main/java/umc/study/converter/MemberConverter.java
@@ -2,8 +2,8 @@ package umc.study.converter;
 
 import umc.study.domain.Member;
 import umc.study.domain.enums.Gender;
-import umc.study.dto.request.MemberRequestDTO;
-import umc.study.dto.response.MemberResponseDTO;
+import umc.study.web.dto.request.MemberRequestDTO;
+import umc.study.web.dto.response.MemberResponseDTO;
 
 import java.time.LocalDateTime;
 import java.util.ArrayList;
