--- src/main/java/umc/study/converter/TempConverter.java
@@ -1,18 +1,0 @@
-package umc.study.converter;
-
-import umc.study.web.dto.TempResponse;
-
-public class TempConverter {
-
-    public static TempResponse.TempTestDTO toTempTestDTO(){
-        return TempResponse.TempTestDTO.builder()
-                .testString("This is Test!")
-                .build();
-    }
-
-    public static TempResponse.TempExceptionDTO toTempExceptionDTO(Integer flag){
-        return TempResponse.TempExceptionDTO.builder()
-                .flag(flag)
-                .build();
-    }
-}
