--- src/main/java/umc/study/service/TempService/TempCommandService.java
@@ -1,4 +1,0 @@
-package umc.study.service.TempService;
-
-public interface TempCommandService {
-}
