--- src/main/java/umc/study/service/TempService/TempCommandServiceImpl.java
@@ -1,4 +1,0 @@
-package umc.study.service.TempService;
-
-public class TempCommandServiceImpl {
-}
