--- src/main/java/umc/study/service/TempService/TempQueryService.java
@@ -1,6 +1,0 @@
-package umc.study.service.TempService;
-
-public interface TempQueryService {
-
-    void CheckFlag(Integer flag);
-}
