--- src/main/java/umc/study/service/TempService/TempQueryServiceImpl.java
@@ -1,17 +1,0 @@
-package umc.study.service.TempService;
-
-import lombok.RequiredArgsConstructor;
-import org.springframework.stereotype.Service;
-import umc.study.apiPayload.code.exception.handler.TempHandler;
-import umc.study.apiPayload.code.status.ErrorStatus;
-
-@Service
-@RequiredArgsConstructor
-public class TempQueryServiceImpl implements TempQueryService{
-
-    @Override
-    public void CheckFlag(Integer flag) {
-        if (flag == 1)
-            throw new TempHandler(ErrorStatus.TEMP_EXCEPTION);
-    }
-}
