--- src/main/java/umc/study/web/controller/TempRestController.java
@@ -1,26 +1,0 @@
-package umc.study.web.controller;
-
-import lombok.RequiredArgsConstructor;
-import org.springframework.web.bind.annotation.GetMapping;
-import org.springframework.web.bind.annotation.RequestMapping;
-import org.springframework.web.bind.annotation.RequestParam;
-import umc.study.apiPayload.ApiResponse;
-import umc.study.converter.TempConverter;
-import umc.study.web.dto.TempResponse;
-
-@RequestMapping("/temp")
-@RequiredArgsConstructor
-public class TempRestController {
-
-    @GetMapping("/test")
-    public ApiResponse<TempResponse.TempTestDTO> testAPI(){
-
-        return ApiResponse.onSuccess(TempConverter.toTempTestDTO());
-    }
-
-    @GetMapping("/exception")
-    public ApiResponse<TempResponse.TempExceptionDTO> exceptionAPI(@RequestParam Integer flag){
-
-        return null;
-    }
-}
