--- src/main/java/umc/study/web/dto/TempResponse.java
@@ -1,25 +1,0 @@
-package umc.study.web.dto;
-
-import lombok.AllArgsConstructor;
-import lombok.Builder;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-
-public class TempResponse {
-
-    @Builder
-    @Getter
-    @NoArgsConstructor
-    @AllArgsConstructor
-    public static class TempTestDTO{
-        String testString;
-    }
-
-    @Builder
-    @Getter
-    @NoArgsConstructor
-    @AllArgsConstructor
-    public static class TempExceptionDTO{
-        Integer flag;
-    }
-}
